[JS IR] Copy IR dumps to the build directory if the compiler crashes
Previously in 3fb1096c18f2f1be82823ab7c997c8e6e9a8393b we've implemented
dumping the IR after each phase when running JS IR box tests.

However, because those dumps are saved to a temporary directory,
they can be lost in case we never make it to
the `JsArtifactsDumpHandler` step where files from that temporary
directory are copied to the `out` directory.
This can happen if, for example, the compiler crashes.
In that case, having IR dumps is even more useful, as they can help
investigate the crash.

Here we introduce an additional step to copy those dumps to the `out`
directory if the test crashes.